third_party/WebKit/LayoutTests/svg/as-image/svgview-references-use-counters-expected.txt

PASS window.internals.isUseCounted(document.querySelector('object').contentDocument, SVGSVGElementFragmentSVGView) is false
PASS window.internals.isUseCounted(document.querySelector('object').contentDocument, SVGSVGElementFragmentSVGViewElement) is false
PASS window.internals.isUseCounted(document.querySelector('object').contentDocument, SVGSVGElementFragmentSVGView) is true
PASS window.internals.isUseCounted(document.querySelector('object').contentDocument, SVGSVGElementFragmentSVGViewElement) is false
PASS window.internals.isUseCounted(document.querySelector('object').contentDocument, SVGSVGElementFragmentSVGView) is true
PASS window.internals.isUseCounted(document.querySelector('object').contentDocument, SVGSVGElementFragmentSVGViewElement) is true
PASS successfullyParsed is true

TEST COMPLETE
  
